sstable: split Layout out of reader.go
